Move tiller posthook to a separate file
